Split function/project creation into separate files
This helps isolate language-specific code and will be more important as we add C# class library support. It should have no functional effect.
