+++ src/extension.ts
@@ -13,8 +13,8 @@ import { AzureAccount } from './azure-account.api';
 import { configureDeploymentSource } from './commands/configureDeploymentSource';
 import { copyFunctionUrl } from './commands/copyFunctionUrl';
 import { createChildNode } from './commands/createChildNode';
-import { createFunction } from './commands/createFunction';
-import { createNewProject } from './commands/createNewProject';
+import { createFunction } from './commands/createFunction/createFunction';
+import { createNewProject } from './commands/createNewProject/createNewProject';
 import { deleteNode } from './commands/deleteNode';
 import { deploy } from './commands/deploy';
 import { editAppSetting } from './commands/editAppSetting';
